# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Permy API",
    "description": "接客返信文生成API（本文非保存・privacy-first）",
    "contact": {
      "name": "Permy Support"
    },
    "license": {
      "name": "Private"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/anonymous": {
      "post": {
        "summary": "Auth Anonymous",
        "operationId": "auth_anonymous_api_v1_auth_anonymous_post",
        "parameters": [
          {
            "name": "X-Device-Fingerprint",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Device-Fingerprint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthAnonymousResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/diagnosis": {
      "post": {
        "summary": "Evaluate Diagnosis",
        "operationId": "evaluate_diagnosis_api_v1_me_diagnosis_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/telemetry/events": {
      "post": {
        "summary": "Post Telemetry Events",
        "description": "Telemetryイベントを受信して保存（本文ゼロ厳守）\n\n- サーバでserver_time_utc, hour_bucket_utc, dow_utc, user_id_hashを付与\n- 会話本文/生成本文は絶対に含めない\n- event_data には各イベント固有のフィールドをJSON保存",
        "operationId": "post_telemetry_events_api_v1_telemetry_events_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppOpenedEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "const": "app_opened",
            "title": "Event Name",
            "default": "app_opened"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "os": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Os"
          },
          "device_class": {
            "type": "string",
            "enum": [
              "phone",
              "tablet",
              "unknown"
            ],
            "title": "Device Class",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "os"
        ],
        "title": "AppOpenedEvent"
      },
      "AuthAnonymousResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "access_token"
        ],
        "title": "AuthAnonymousResponse"
      },
...
      "CandidateCopiedEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "const": "candidate_copied",
            "title": "Event Name",
            "default": "candidate_copied"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "os": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Os"
          },
          "device_class": {
            "type": "string",
            "enum": [
              "phone",
              "tablet",
              "unknown"
            ],
            "title": "Device Class",
            "default": "unknown"
          },
          "candidate_id": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C"
            ],
            "title": "Candidate Id"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "os",
          "candidate_id"
        ],
        "title": "CandidateCopiedEvent"
      },
...
      "DiagnosisAnswer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "choice_id": {
            "type": "string",
            "title": "Choice Id"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "choice_id"
        ],
        "title": "DiagnosisAnswer"
      },
      "DiagnosisRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/DiagnosisAnswer"
            },
            "type": "array",
            "maxItems": 7,
            "minItems": 7,
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "DiagnosisRequest"
      },
      "DiagnosisResponse": {
        "properties": {
          "persona_version": {
            "type": "integer",
            "title": "Persona Version"
          },
          "true_self_type": {
            "type": "string",
            "enum": [
              "Stability",
              "Independence",
              "Approval",
              "Realism",
              "Romance"
            ],
            "title": "True Self Type"
          },
          "night_self_type": {
            "type": "string",
            "enum": [
              "VisitPush",
              "Heal",
              "LittleDevil",
              "BigClient",
              "Balance"
            ],
            "title": "Night Self Type"
          },
          "persona_goal_primary": {
            "type": "string",
            "enum": [
              "next_visit",
              "relationship_keep",
              "long_term_growth",
              "firefighting",
              "special_distance"
            ],
            "title": "Persona Goal Primary"
          },
          "persona_goal_secondary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Goal Secondary"
          },
          "style_assertiveness": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Style Assertiveness"
          },
          "style_warmth": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Style Warmth"
          },
          "style_risk_guard": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Style Risk Guard"
          }
        },
        "type": "object",
        "required": [
          "persona_version",
          "true_self_type",
          "night_self_type",
          "persona_goal_primary",
          "style_assertiveness",
          "style_warmth",
          "style_risk_guard"
        ],
        "title": "DiagnosisResponse"
      },
...
      "GenerateFailedEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "const": "generate_failed",
            "title": "Event Name",
            "default": "generate_failed"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "os": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Os"
          },
          "device_class": {
            "type": "string",
            "enum": [
              "phone",
              "tablet",
              "unknown"
            ],
            "title": "Device Class",
            "default": "unknown"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "os",
          "error_code"
        ],
        "title": "GenerateFailedEvent"
      },
...
      "GenerateRequestedEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "const": "generate_requested",
            "title": "Event Name",
            "default": "generate_requested"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "os": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Os"
          },
          "device_class": {
            "type": "string",
            "enum": [
              "phone",
              "tablet",
              "unknown"
            ],
            "title": "Device Class",
            "default": "unknown"
          },
          "daily_used": {
            "type": "integer",
            "title": "Daily Used"
          },
          "daily_remaining": {
            "type": "integer",
            "title": "Daily Remaining"
          },
          "has_ng_setting": {
            "type": "boolean",
            "title": "Has Ng Setting"
          },
          "persona_version": {
            "type": "integer",
            "title": "Persona Version"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "os",
          "daily_used",
          "daily_remaining",
          "has_ng_setting",
          "persona_version"
        ],
        "title": "GenerateRequestedEvent"
      },
...
      "GenerateSucceededEvent": {
        "properties": {
          "event_name": {
            "type": "string",
            "const": "generate_succeeded",
            "title": "Event Name",
            "default": "generate_succeeded"
          },
          "app_version": {
            "type": "string",
            "title": "App Version"
          },
          "os": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "title": "Os"
          },
          "device_class": {
            "type": "string",
            "enum": [
              "phone",
              "tablet",
              "unknown"
            ],
            "title": "Device Class",
            "default": "unknown"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "ng_gate_triggered": {
            "type": "boolean",
            "title": "Ng Gate Triggered"
          },
          "followup_returned": {
            "type": "boolean",
            "title": "Followup Returned"
          }
        },
        "type": "object",
        "required": [
          "app_version",
          "os",
          "latency_ms",
          "ng_gate_triggered",
          "followup_returned"
        ],
        "title": "GenerateSucceededEvent"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryEventRequest": {
        "properties": {
          "events": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GenerateRequestedEvent"
                },
                {
                  "$ref": "#/components/schemas/GenerateSucceededEvent"
                },
                {
                  "$ref": "#/components/schemas/GenerateFailedEvent"
                },
                {
                  "$ref": "#/components/schemas/CandidateCopiedEvent"
                },
                {
                  "$ref": "#/components/schemas/AppOpenedEvent"
                }
              ]
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "TelemetryEventRequest",
        "description": "POST /api/v1/telemetry/events のリクエスト"
      },
      "TelemetryEventResponse": {
        "properties": {
          "received": {
            "type": "integer",
            "title": "Received"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "received"
        ],
        "title": "TelemetryEventResponse",
        "description": "POST /api/v1/telemetry/events のレスポンス"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}